requirements.txt
matplotlib>=3.10.7
numpy>=2.3.4
imageio-ffmpeg>=0.6.0
commentjson>=0.9.0
questionary>=2.1.1

# Follow the install guide for CuPy:
# https://docs.cupy.dev/en/stable/install.html